Refactor: update tool import
