+++ src/program.ts
@@ -24,7 +24,7 @@ import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
 import { SSEServerTransport } from '@modelcontextprotocol/sdk/server/sse.js';
 
 
-import { createServer } from './index';
+import { createServer } from '../packages/mcp/server/index';
 import { ServerList } from './server';
 
 import type { LaunchOptions } from 'playwright';
